Improve naming scheme for decompilation and decompile blocks
In order to prevent clashing with Clojure's name (eg. `loop` in WASM and
`loop` in Clojure), decompilation function which directly map to a
non-terminal WASM symbol are suffixed with a '. As such, they are
recognizable and tend to mimick what they decompile (eg. the `block'`
function indeed decompiles a `'block` into a list `(block ...)`).