Decompile code into a dedicated key of the context
Decompiled data, which is already similar to WAT, will be further
processed to approach WAT even closer. It is important to first
decompile into a state that is reminiscent of the structure of the
binary file, but it is also useful to easily discard all that
information when it becomes useless.