Make raw decompilation already WAT friendly
Up to now, there was a distinction between raw decompilation function
which decompile WASM into Clojure datastructures that directly mimicks
the binary format and newer decompilation functions which decompiles
almost directly to WAT. The distinction between both categories was
mainly stemming from a need for performance.

However, remaining that close to the binary format is not really
necessary. By decompiling to something that is already closer to WAT,
we can obtain something that is still fairly close to the binary format
while being ready for transformation into WAT without double parsing
some sections.

The main notable change is that the code section is not decompiled, only
split into views (one per function). That is because code decompilation
might involve more aspects such as resolving WASM indices to symbolic
identifiers for human-readibility.