Track symbolic id param substitution at decompilation
At decompilation, function parameters are being given symbolic
identifiers based on their index so that the code will be eventually
more human-readable.

Now, a map of param id -> param index is kept in the metadata of a
function. This will simplify implementation of future features (eg.
substitution of indices for ids in codesec, type-checking, etc).