Decompile typesec and importsec into IR
Up to now, the decompiled format was a bit shaky. Should it remain close
to the binary format? Closer to WAT?

Naturally, it all depends on the intended purpose. Working on this
problem shows that its best to resolve to an lightweight IR
(intermediary representation) in-between. Staying close to the binary
format is not practical for further work and if performance for very
specific tasks is needed (eg. validating right away), they can be
implemented somewhat easily by levering the codebase. Staying close to
WAT is also impractical as working exclusively with lists is cumbersome.
An IR with combines both maps, vectors, and WAT-looking lists seems to
be the best approach in the sense it is the most versatile one.