Decompile typesec into a proper section and favor bin over sym
Previously, the type section was decompiled just so that other sections
referencing types could resolve type indices. However, the new direction
is to maintain a genuine type section and stop resolving type indices.
This will have a few advantages later, especially when writing WASM
files.

Also, this commit introduces keeping binary values instead of converting
them to symbols (eg. keeping a valtype as a byte value instead of
resolving to a human-readable symbol). The purpose of the IR is to have
a flexible datastructures which represents a WASM file. Resolving to
symbols would only serve human-readability at the cost of time and
space. When writing algorithms operating over the IR, the code remains
easy to understand since all those binary values of interest are or will
be interned in the `helins.wasm.bin` namespace.