Add counting typesec
In order to achieve great performance and ease of mind when writing a
WASM files, all byte counts will be counted prior to writing. This
allows for allocating exactly the right amount of byte and writing
without defensive boundary checking. It even opens the possibility for
fast multithreaded writing using mmap.